+++ CHANGELOG.md
@@ -10,6 +10,8 @@
 - Added option to hide the nametag of your spawned summoning mobs
 - Added option to mark the own summoning mobs in green
 - Added highlight for the different ashfang blazes in their respective color
+- Added option to hide all the particles around the ashfang boss
+- Added option to hide the name of full health blazes around ashfang (only useful when highlight blazes is enabled)
 
 ### Minor Changes
 - Optimizing the highlight block size for minions, blazing souls and gravity orbs